address reviewers commetns
